Allow the `accession` sub-command to assign multiple IDs
Assigning accessions to three files, giving them the IDs
"MYID001", "MYID002", and "MYID003":

./sda-admin accession 'MYID%03d' 1 file1 file2 file3

Assigning accession IDs to all files in or beneath the path
"project/data", starting at 123.  The first two accession
IDs will be "PROJ00000123" and "PROJ00000124".  The ordering
in which accession IDs are assigned will depend on the
ordering of the files in the RabbitMQ queue.

./sda-admin accession 'PROJ%08' 123 project/data/