Add `-a` option to some sub-commands
The `ingest`, `accession`, and `dataset` sub-commands gain a `-a`
command line option.  This option is valid when the sub-command is used
for listing the available filenames in a given queue, and makes the
command list the filenames in _all_ buckets.

Note that the output format changes when `-a` is used as the filenames
are prefixed with the bucket thathey reside in.