Don't allow upload targed directory path to start with `./`
There has been a change in `sda-cli` or in MinIO such that if we
specify the a target directory with `-targetDir`, that directory
path may not contain a dot component.  We previously used a dot
at the start (as in `./some/dir/path`) but this commit removes
any `./` substring from the start of the path.